Almost fix red animals
The fix was to extract the materials from the .blend, then add a custom shader to be able to use Vertex Colors, and apply it to the extracted materials.

Without URP, this would have been enough. Instead, I had to try to readapt the shader, because it didn't work well with URP. There are still compilation errors, and the resulting models now lack (at least) lighting and shadows